requirements.txt
fastapi
httpx
beautifulsoup4
playwright
python-dotenv
groq
pydantic